DEVPROD-16721: switch role used for writes to new bucket
This commit adjusts the role we use to write to our new bucket. The previous
role will only work when we separate out Atlas CLI from Mongo CLI's
infrastructure. Because the two share a common bucket for now, I've
swapped over to the appropriate role with access to that bucket.